backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database and ORM
sqlmodel>=0.0.14
sqlalchemy>=2.0.23
alembic>=1.13.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Database
chromadb==0.5.23

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.1
pytesseract>=0.3.10
Pillow>=10.1.0

# AI and Embeddings
openai>=1.3.0
anthropic>=0.7.0
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0

# File Handling and Validation
aiofiles>=23.2.0
python-magic>=0.4.27

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# WebSocket support
websockets>=11.0.3

# Logging and Monitoring
structlog>=23.2.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI

# Type checking and linting
mypy>=1.7.0
ruff>=0.1.6
black>=23.11.0

# Production
gunicorn>=21.2.0

# Optional: For production PostgreSQL support
# psycopg2-binary>=2.9.7

# Optional: For Redis caching
# redis>=5.0.0

# Optional: For email notifications
# fastapi-mail>=1.4.1

# Optional: For file type detection
# filetype>=1.2.0